Swanky new Camera in J3DScene, pan and zoom in J2DScene and loads of little changes
git-svn-id: svn+ssh://ask.diku.dk/usr/local/projects/disk03/proteins/repo/ProGAL/Java@765 53b6c913-c175-4e75-a8ba-878503ff61f7

+++ src/ProGAL/geom2d/viewer/TextShape.java
@@ -5,16 +5,16 @@ import ProGAL.geom2d.Shape;
 
 public class TextShape implements Shape {
 	private String text;
-	private float height;
+	private double height;
 	private Point pos;
 	
-	public TextShape(String text, Point pos, float height){
+	public TextShape(String text, Point pos, double height){
 		this.text = text;
 		this.height = height;
 		this.pos = pos;
 	}
 	
-	public float getHeight(){ return height; }
+	public double getHeight(){ return height; }
 	public String getText(){ return text; }
 	public Point getPos(){ return pos; }
